pos-backend/requirements.txt
pytz
Django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-spectacular==0.27.2
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
redis==5.0.7
psycopg2-binary==2.9.9   # if Postgres
python-dotenv==1.0.1
requests==2.32.5
urllib3<2.0  # Pin to v1.x for LibreSSL compatibility on macOS
